+++ api/jest.config.js
@@ -2,4 +2,7 @@ module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'node',
   rootDir: 'src',
+  verbose: true,
+  collectCoverage: true,
+  coverageDirectory: '../coverage',
 };