+++ app/src/pages/Projects.tsx
@@ -1,22 +1,50 @@
 import * as React from "react";
-import { Table } from "react-bootstrap";
+import { useState } from "react";
+import { Table, Glyphicon } from "react-bootstrap";
 import { Link } from "react-router-dom";
 
 import { useProjects } from "../api";
+import { compareProjectSummary } from "../summarize";
 
 const style = { padding: '10px' };
 
+interface ColumnHeaderProps {
+    active: string;
+    skey: string;
+    onClick(key: string): void;
+    title: string;
+}
+
+const ColumnHeader = ({ active, skey, onClick, title }: ColumnHeaderProps) => (
+    <th>
+        <span title="Click to sort on" style={
+            active !== skey ? { cursor: 'pointer' } : {}
+        } onClick={() => active !== skey && onClick(skey)}>{title}</span>
+        {active === skey && <Glyphicon glyph="sort-by-attributes-alt" />}
+    </th>
+)
+
 export function Projects() {
     const projects = useProjects();
+    const [sortkey, setSortKey] = useState('met_id');
+
+    const sortOn = (key: string) => {
+        // TODO reverse sort when sorted column is clicked again
+        const data = [...projects.data];
+        data.sort(compareProjectSummary(key));
+        projects.setData({data});
+        setSortKey(key);
+    }
+
     if (projects.loading) {
         return <span>Loading ...</span>;
     }
     if (projects.error) {
-        return <span>Error: {projects.error}</span>
+        return <span>Error: {projects.error.message}</span>
     }
     const rows = projects.data.map(d => (
         <tr key={d._id}>
-            <td><Link to={`/projects/${d._id}`}>{ d.GNPSMassIVE_ID ? d.GNPSMassIVE_ID : d.metabolights_study_id }</Link></td>
+            <td><Link to={`/projects/${d._id}`}>{d.GNPSMassIVE_ID ? d.GNPSMassIVE_ID : d.metabolights_study_id}</Link></td>
             <td>{d.PI_name}</td>
             <td>{d.submitters}</td>
             <td>{d.nr_genomes}</td>
@@ -33,15 +61,15 @@ export function Projects() {
             <Table>
                 <thead>
                     <tr>
-                        <th>Metabolomics project identifier</th>
-                        <th>Principal investigator</th>
-                        <th>Submitter(s)</th>
-                        <th>Nr of (meta)genomes</th>
-                        <th>Nr of growth conditions</th>
-                        <th>Nr of extraction methods</th>
-                        <th>Nr of instrumentation methods</th>
-                        <th>Nr of links between genome and metabolome samples</th>
-                        <th>Nr of links between gene clusters and MS2 spectra</th>
+                        <ColumnHeader skey="met_id" active={sortkey} onClick={sortOn} title="Metabolomics project identifier" />
+                        <ColumnHeader skey="PI_name" active={sortkey} onClick={sortOn} title="Principal investigator" />
+                        <ColumnHeader skey="submitters" active={sortkey} onClick={sortOn} title="Submitter(s)" />
+                        <ColumnHeader skey="nr_genomes" active={sortkey} onClick={sortOn} title="Nr of (meta)genomes" />
+                        <ColumnHeader skey="nr_growth_conditions" active={sortkey} onClick={sortOn} title="Nr of growth conditions" />
+                        <ColumnHeader skey="nr_extraction_methods" active={sortkey} onClick={sortOn} title="Nr of extraction methods" />
+                        <ColumnHeader skey="nr_instrumentation_methods" active={sortkey} onClick={sortOn} title="Nr of instrumentation methods" />
+                        <ColumnHeader skey="nr_genome_metabolmics_links" active={sortkey} onClick={sortOn} title="Nr of links between genome and metabolome samples" />
+                        <ColumnHeader skey="nr_genecluster_mspectra_links" active={sortkey} onClick={sortOn} title="Nr of links between gene clusters and MS2 spectra" />
                     </tr>
                 </thead>
                 <tbody>